requirements.txt
fastapi[standard]
sqlalchemy
psycopg2
python-dotenv